Mostly Finished Reducers
Ran into a blocker when trying to make combined reducers into 2 reducers. Both reducers would use the same data in state, so it seemed in this case you would just make a single index.js reducer and put all the references to the action creators into that one reducer. I'm not sure if it would be possible/suggested to make 2 reducers in this case.