Added isLoading state BUGGED
Added an isLoading state in my top-level component to try to add a loading spinner to prevent a bug caused by trying to route to modifylist before the CDM fires for the friendslist component. I would do this by removing the navlink until the CDM is done and replacing it with a spinner. My routes now seem to be broken, going to try moving the logic for the NavLink into the view component instead